Products summary, No-Image Icon
Added products summary which keeps track of quantities that were updated, blocks buttons if there are no products to update and blocks Mark As Sold button if there exists a product which quantity will be larger than previously (because when you sell an item it is logical that its quantity cant go up), backend functionality that differentiates these buttons to be implemented in the future. Added a default image which displays if product has no image. Removed Hat.svg from components folder